refactor/fix pool tests
